requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
future==0.17.1
idna==2.7
nose==1.3.7
pkg-resources==0.0.0
pycparser==2.19
python-telegram-bot==11.1.0
six==1.11.0
